Convert old Harness file to typescript to make it compile
